aider: adom.py now uses ADOM_PATH when launching the game.
# Aider chat conversation:

USER: adom.py ignores ADOM_PATH

ASSISTANT: The issue seems to be that the `ADOM_PATH` environment variable is not being used when launching the ADOM game in the `main()` function of `adom.py`.

To fix this, we need to modify `adom.py` to use `ADOM_PATH` when launching the game.

Here is the *SEARCH/REPLACE* block:

adom.py
```python
<<<<<<< SEARCH
    adom_args = [adom_path]
=======
    adom_args = [adom_path if adom_path else 'adom']
>>>>>>> REPLACE
```

This change will ensure that the `ADOM_PATH` environment variable is used when it is set. If it is not set, the default value 'adom' will be used.